Module halo2_debug::display

Structs

ExprDisp

Wrapper type over Expression that implements Display with nice output. The formatting of the Expression::Variable case is parametrized with the second field, which take as auxiliary value the third field. Use the constructor expr_disp to format variables using their Display implementation. Use the constructor expr_disp_names for an Expression with V: VarMid to format column queries according to their string names.

FDisp

Wrapper type over PrimeField that implements Display with nice output.

LookupArgDisp

Wrapper type over lookup::Argument that implements Display with nice output. The formatting of the Expression::Variable case is parametrized with the second field, which take as auxiliary value the third field. Use the constructor lookup_arg_disp to format variables using their Display implementation. Use the constructor lookup_arg_disp_names for a lookup of Expression with V: VarMid that formats column queries according to their string names.

ShuffleArgDisp

Wrapper type over shuffle::Argument that implements Display with nice output. The formatting of the Expression::Variable case is parametrized with the second field, which take as auxiliary value the third field. Use the constructor shuffle_arg_disp to format variables using their Display implementation. Use the constructor shuffle_arg_disp_names for a shuffle of Expression with V: VarMid that formats column queries according to their string names.

Functions

expr_disp

ExprDisp constructor that formats viariables using their Display implementation.

expr_disp_names

ExprDisp constructor for an Expression with V: VarMid that formats column queries according to their string names.

lookup_arg_disp

LookupArgDisp constructor that formats viariables using their Display implementation.

lookup_arg_disp_names

LookupArgDisp constructor for a lookup of Expression with V: VarMid that formats column queries according to their string names.

shuffle_arg_disp

ShuffleArgDisp constructor that formats viariables using their Display implementation.

shuffle_arg_disp_names

ShuffleArgDisp constructor for a shuffle of Expression with V: VarMid that formats column queries according to their string names.